backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
pdfplumber
unidecode
python-dotenv
pymupdf
